FIX reconnection with PAT or PMT with length != 1
When the length of the PAT or PMT tables is more than one single packet, the processing fails when checking with the original table. The behaviour is then that the input it's reconnected with any new table.
This patch fixes this issue and continues processing until the complete new table is available; and only after this it checks it with the previous table. Then it only reconnects when a real change is detected.